externalize docker-compose prompts
