+++ generators/base-workspaces/generator.ts
@@ -17,12 +17,7 @@
  * limitations under the License.
  */
 
-import { readdir } from 'fs/promises';
-import { existsSync } from 'fs';
-import chalk from 'chalk';
-
 import BaseGenerator from '../base/index.js';
-import { YO_RC_FILE } from '../generator-constants.js';
 import { GENERATOR_BOOTSTRAP_APPLICATION } from '../generator-list.js';
 import { normalizePathEnd } from '../../lib/utils/index.js';
 import { CONTEXT_DATA_APPLICATION_KEY } from '../base-simple-application/support/index.js';
@@ -76,6 +71,7 @@ export default abstract class BaseWorkspacesGenerator<
 
   static PREPARING_WORKSPACES = BaseGenerator.asPriority(PREPARING_WORKSPACES);
 
+  customWorkspacesConfig?: boolean;
   appsFolders?: string[];
   directoryPath!: string;
 
@@ -109,7 +105,7 @@ export default abstract class BaseWorkspacesGenerator<
     });
   }
 
-  get sharedWorkspaces(): any {
+  get #sharedWorkspaces(): any {
     return this.getContextData(CONTEXT_DATA_WORKSPACES_KEY, { factory: () => ({}) });
   }
 
@@ -121,56 +117,6 @@ export default abstract class BaseWorkspacesGenerator<
 
   protected configureWorkspacesConfig() {
     this.jhipsterConfig.directoryPath = normalizePathEnd(this.jhipsterConfig.directoryPath ?? './');
-  }
-
-  protected async askForWorkspacesConfig() {
-    let appsFolders;
-    await this.prompt(
-      [
-        {
-          type: 'input',
-          name: 'directoryPath',
-          message: 'Enter the root directory where your applications are located',
-          default: '../',
-          validate: async input => {
-            const path = this.destinationPath(input);
-            if (existsSync(path)) {
-              const applications = await this.findApplicationFolders(path);
-              return applications.length === 0 ? `No application found in ${path}` : true;
-            }
-            return `${path} is not a directory or doesn't exist`;
-          },
-        },
-        {
-          type: 'checkbox',
-          name: 'appsFolders',
-          when: async answers => {
-            const directoryPath = answers.directoryPath;
-            appsFolders = (await this.findApplicationFolders(directoryPath)).filter(
-              app => app !== 'jhipster-registry' && app !== 'registry',
-            );
-            this.log.log(chalk.green(`${appsFolders.length} applications found at ${this.destinationPath(directoryPath)}\n`));
-            return true;
-          },
-          message: 'Which applications do you want to include in your configuration?',
-          choices: () => appsFolders,
-          default: () => appsFolders,
-          validate: input => (input.length === 0 ? 'Please choose at least one application' : true),
-        },
-      ],
-      this.config,
-    );
-  }
-
-  protected async findApplicationFolders(directoryPath = this.directoryPath ?? '.') {
-    return (await readdir(this.destinationPath(directoryPath), { withFileTypes: true }))
-      .filter(dirent => dirent.isDirectory())
-      .map(dirent => dirent.name)
-      .filter(
-        folder =>
-          existsSync(this.destinationPath(directoryPath, folder, 'package.json')) &&
-          existsSync(this.destinationPath(directoryPath, folder, YO_RC_FILE)),
-      );
   }
 
   private resolveApplicationFolders({
@@ -213,7 +159,7 @@ export default abstract class BaseWorkspacesGenerator<
     return [
       {
         ...first,
-        workspaces: this.sharedWorkspaces,
+        workspaces: this.#sharedWorkspaces,
         deployment: this.context,
         applications: this.#applications,
       },
